% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{check_function2}
\alias{check_function2}
\title{Example high-level SCT for checking a function}
\usage{
check_function2(state, name, index, args)
}
\arguments{
\item{state}{state to start from}

\item{name}{passed to check_function}

\item{index}{passed to check_function}

\item{args}{an array of argument names, passed to check_args, then check_equal.}
}
\description{
Example high-level SCT for checking a function
}

