% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-description.R
\name{check_has_desc_element}
\alias{check_has_desc_element}
\alias{check_desc_element_matches}
\alias{check_desc_version}
\alias{check_desc_date}
\alias{check_desc_authors_at_r}
\title{Check the DESCRIPTION fields}
\usage{
check_has_desc_element(state, element, missing_msg = NULL, append = TRUE)

check_desc_element_matches(state, element, regex, fixed = FALSE, times = 1L,
  not_typed_msg = NULL, append = TRUE)

check_desc_version(state, expected, bad_format_msg = NULL,
  incorrect_msg = NULL, append = TRUE)

check_desc_date(state, expected = Sys.Date(), bad_format_msg = NULL,
  incorrect_msg = NULL, append = TRUE)

check_desc_authors_at_r(state, expected, bad_format_msg = NULL,
  incorrect_msg = NULL, append = TRUE)
}
\arguments{
\item{state}{The state of the exercise, as returned from \code{\link{parse_desc}}.}

\item{element}{String naming the element of the roxygen block to check.}

\item{missing_msg}{Optional string. Used to override the feedback message
in the event of failure.}

\item{append}{For compatibility; currently unused.}

\item{regex}{String providing a regular expression for the solution code to
match. See \code{testwhat}'s \code{check_code} function.}

\item{fixed}{Logical. If \code{TRUE}, regex is treated as a fixed string, not
a regular expression. See \code{testwhat}'s \code{check_code} function.}

\item{times}{Positive integer. Denotes the number of times the string in
\code{regex} should be matched.}

\item{not_typed_msg}{Optional string. Used to override the feedback message
in the event of failure.}

\item{expected}{The expected value.}

\item{bad_format_msg}{Optional string. Used to override the feedback message
in the event of failure.
incorrect_msg}

\item{incorrect_msg}{Optional string. Used to override the feedback message
in the event of failure.}
}
\value{
This function is invoked for the side effect of registering feedback
in the event of a failed test. See \code{\link{check_that}} for details of
the return value and feedback mechanism.
\code{check_has_desc_element} checks that the \code{element} element of the
DESCRIPTION is present.
\code{check_desc_element_equals} checks that the \code{element} element of
the DESCRIPTION is equal to the value in the solution code.
\code{check_desc_element_matches} checks that the \code{element} element of
the DESCRIPTION matches a regular expression or string.
}
\description{
Check that the DESCRIPTION fields provided by the student are correct.
}
\note{
None of these functions use solution code. In order to facilitate this,
the design of \code{check_desc_version},  \code{check_desc_date}, and
 \code{check_desc_authors_at_r} differs from other testwhat check functions
 in that they require an \code{expected} argument, which takes the expected
 value. This is coerced into a \code{package_version}, \code{Date}, or
 \code{person} object.
}
\examples{
\dontrun{
  # Always begin by calling parse_desc() on the exercise state.
  ex() \%>\% parse_desc() \%>\% {
    check_has_desc_element(., 'title')
    check_roxy_element_matches(., 'description', 'data.+manipulation')
    check_desc_version(., '1.0.0')
    check_desc_date(., '1919-10-18')
    check_desc_authors_at_r(.,
      'Richie Cotton <richie@datacamp.com> [cre, aut]'
    )
  }
}

}
