% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-lint.R
\name{check_is_lint_free}
\alias{check_is_lint_free}
\title{Check that there is no lint}
\usage{
check_is_lint_free(state)
}
\arguments{
\item{state}{The state of the exercise, as returned from 
\code{\link{make_lint_state}}.}
}
\value{
The child state.
}
\description{
Check that the student's code is lint-free
}
