% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-lavaan.R
\name{check_lavaan_pattern}
\alias{check_lavaan_pattern}
\title{Run \code{grepl} on lavaan model strings.}
\usage{
check_lavaan_pattern(state, patt, not_string_msg = NULL, incorrect_msg,
  append = TRUE)
}
\arguments{
\item{state}{state to start from (should be state produced by
\code{check_object}).}

\item{patt}{pattern to look for in the lavaan model string that was targeted
with \code{check_object}.}

\item{not_string_msg}{if specified, this overrides the automatically
generated message in case the targetted object is not a string.}

\item{incorrect_msg}{feedback message if calling the function specified in
func for if the test fails}

\item{append}{whether or not to append the feedback messages to earlier
messages that describe the object.}
}
\description{
If the pattern is not found, the function throws an error message (that you
have to specify explicitly).
}
\examples{
\dontrun{

# solution model
mod <- 'a =~ x1 + x2'
# student model

# scts that pass
ex() \%>\%
  check_object("model") \%>\%
  check_lavaan_pattern("a")

ex() \%>\%
  check_object("model") \%>\%
  check_lavaan_pattern("a\\\\s+=~\\\\s+x1")

# scts that fail
ex() \%>\%
  check_object("model") \%>\%
  check_lavaan_pattern("x2")

ex() \%>\%
  check_object("model") \%>\%
  check_lavaan_pattern("x1\\\\s\\\\+\\\\s+x2")
}

}
