% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{check_object2}
\alias{check_object2}
\title{Example high-level SCT for checking object equality}
\usage{
check_object2(state, names)
}
\arguments{
\item{state}{state to start from}

\item{names}{one or more object names to test for equality}
}
\description{
Example high-level SCT for checking object equality
}

