% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-rcpp.R
\name{check_rcpp}
\alias{check_rcpp}
\alias{check_embedded_r}
\alias{check_rcpp}
\alias{check_cpp}
\title{Parse cpp files with embedded R}
\usage{
check_embedded_r(state)

check_cpp(state)
}
\arguments{
\item{state}{An exercise state, as returned by \code{ex()}.}
}
\value{
A child state.
}
\description{
Extracts the embedded R and Cpp portions from a cpp file.
}
