% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-equality.R
\name{equality}
\alias{equality}
\alias{are_classes_equal}
\alias{are_lengths_equal}
\alias{are_dims_equal}
\alias{are_igraphs_equal}
\alias{are_corpora_equal}
\title{Custom equality checks}
\usage{
are_classes_equal(x, y)

are_lengths_equal(x, y)

are_dims_equal(x, y)

are_igraphs_equal(x, y)

are_corpora_equal(x, y)
}
\arguments{
\item{x}{The student's object.}

\item{y}{The solution object.}
}
\value{
\code{TRUE} or \code{FALSE}, depending on whether or not the two
objects are considered the same. See details.
}
\description{
Custom equality checks to make it easier to check weird objects.
}
\details{
The functions performs the following checks.
\describe{
\item{\code{are_classes_equal()}}{Identical classes.}
\item{\code{are_lengths_equal()}}{Identical length.}
\item{\code{are_dims_equal()}}{Identical dimensions. For objects with
\code{NULL} dimension, their length is used.}
\item{\code{are_igraphs_equal()}}{\code{igraphs} with identical edges and
directedness.}
\item{\code{are_corpora_equal()}}{Corpora with identical contents.}
}
}
