% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-lint.R
\name{lint_from_code}
\alias{lint_from_code}
\title{Write code lines to file and get lint msgs}
\usage{
lint_from_code(lines)
}
\arguments{
\item{lines}{A character vector of R code.}
}
\value{
A character vector of lint messages.
}
\description{
Writes a character vector of R code to a file and runs 
\code{\link[lintr]{lintr}} on it.
}
