% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-lint.R
\name{make_lint_state}
\alias{make_lint_state}
\title{Make a lint exercise state}
\usage{
make_lint_state(state)
}
\arguments{
\item{state}{An exercise state, as returned by \code{ex()}.}
}
\value{
A child state.
}
\description{
Convert the parse data to lint messages.
}
