% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-roxygen.R
\name{parse_desc}
\alias{parse_desc}
\title{Parse DESCRIPTION}
\usage{
parse_desc(state)
}
\arguments{
\item{state}{An exercise state, as returned by \code{ex()}.}
}
\value{
A child state.
}
\description{
Parses a package DESCRIPTION file and updates the state.
}
\details{
The function extracts the DESCRIPTION fields from the state then
parses them.
}
