% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_roxy}
\alias{parse_roxy}
\title{Parse roxygen2 comments}
\usage{
parse_roxy(state)
}
\arguments{
\item{state}{An exercise state, as returned by \code{ex()}.}
}
\value{
A child state.
}
\description{
Parses roxygen2 comments and updates the state.
}
\details{
The function extracts the roxygen2 comments from the state then
parses them.
}
